requirements.txt
appdirs==1.4.4
apturl==0.5.2
autopep8==1.5.7
bios==0.1.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debtcollector==2.2.0
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dogpile.cache==1.1.2
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.10
httplib2==0.14.0
idna==2.8
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.1
keyring==18.0.1
keystoneauth1==4.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==1.1.0
msgpack==1.0.2
munch==2.5.0
mysql-connector-python==8.0.23
netaddr==0.8.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openstacksdk==0.55.0
os-service-types==1.7.0
oslo.config==8.5.0
oslo.i18n==5.0.1
oslo.serialization==4.1.0
oslo.utils==4.8.0
oyaml==1.0
packaging==20.9
pbr==5.5.1
peewee==3.14.4
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-keystoneclient==4.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
requestsexceptions==1.4.0
rfc3986==1.4.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
stevedore==3.3.0
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vboxapi==1.0
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
